fix: break up functionality into chunks
